+++ client_utils.h
@@ -1,0 +1,8 @@
+#ifndef CLIENT_UTILS_H
+#define CLIENT_UTILS_H
+
+void *receive_messages(void *socket_desc);
+
+void clean_exit(int sock, pthread_t thread);
+
+#endif
